Rewrite client/backend request correlation
First stab at connecting client requests to backend requests using the
Backend log tag which details which file descriptor was used to handle
the backend request (as well as which backend director and backend
instance was used).

Still a few issues: some times (with many simultaneous requests) an
empty client request is picked up instead of a backend request.